compress/gzip: permit parsing of GZIP files with long header fields
Previously, the GZIP Reader could not have Header.Name or Header.Comment
fields that exceeded 511B, which is problematic for a number of GZIP
files such as OpenBSD packages, which place a signature into the Comment
field. We extend this limit to a larger value of 32KiB, which is comparable
to the memory requirement of DEFLATE (which has a 32KiB window size) and
should not present a DOS security risk to clients and servers that may
need to decompress untrusted GZIP content.

Fixes #14639

Change-Id: Ic1c0743472bf2743343ee5f84d6d75c533f0b15d