net/http/httputil: allow ReverseProxy to call ModifyResponse on failed requests
Previously when RoundTrip returned a non-nil error, the proxy returned a
StatusBadGateway error, instead of first calling ModifyResponse. This
commit first calls ModifyResponse, whether or not the error returned
from RoundTrip is nil. However, it is the user's responsibility to
return an error for a nil response, otherwise ServeHTTP will panic.

A test is added to ensure the following outcomes:
+---------+---------+-------------------------------------+
| Resp    | Error   | Outcome                             |
+---------+---------+-------------------------------------+
| non-nil | non-nil | StatusBadGateway                    |
| non-nil | nil     | StatusOK                            |
| nil     | non-nil | StatusBadGateway                    |
| nil     | nil     | ServeHTTP panics, nil response, EOF |
+---------+---------+-------------------------------------+

Fixes #21255

Change-Id: I5b5bf23a69ae5608f87d4ece756a1b4985ccaa9c