Update patch set 2
Patch Set 2:

One of the main advantages (I find) in subtests is that if the test can fail for several reasons for input `in`, you don't have to repeat the t.Errorf("for input %q, failed because ....", in, ....) noise for every failure, since the context is already in the test name.

But this CL does not do that.

If this simplified the test bodies, then it makes more sense. But it might not have much win in this case. I just skimmed enough to see no changes there.

Patch-set: 2
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>