archive/tar: check for permissible output formats first
The previous logic flow attempted to encode the Header in one
format and if it discovered that it could not it would attempt to
switch to a different format mid-way through. This makes it very
hard to reason about what format will be used in the end and whether
it will even be a valid format.

Instead, we should verify from the start what formats can encode
the given input Header and commit to a given format. This avoids
any hairy logic that tries to switch formats part way through.

Update #9683
Update #12594

Change-Id: I8406ea855dfcb8b478a03a7058ddf8b2b09d46dc