Update patch set 1
Patch Set 1:

(2 comments)

Seems like the CL description should be more clear that this change does not actually pick a format first.  This is the code for picking the format, but then we ignore the format we picked.  Unless I misunderstand.  I assume some future change will actually use the chosen format.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>