Update patch set 1
Patch Set 1:

I feel like the benchmarks are misleading as it seems to really be just measuring the cost of calling delayReader.Read. ReadFrom performs worse only because it has to make more calls since the buffer is smaller. How often is it the case that an io.Reader implementation has such a high per-Read call overhead? In that situation, wouldn't you put an bufio.Reader around it?

Also, while not-intuitive, there are ways to prevent io.CopyBuffer from using ReadFrom and WriteTo methods. See https://github.com/golang/go/issues/16474, which has some discussion about the non-performance of io.CopyBuffer is situations like this.

Patch-set: 1
CC: Gerrit User 9735 <9735@62eb7196-b449-3ce5-99f1-c037f21e1705>