Update patch set 2
Patch Set 2:

Tried this out locally, but not seeing which part of the UI actually changed.

My immediate thought is that it would make the start times and the bar on top turn into wall time, which I wouldn't be a fan of. Having everything shifted to zero makes it easier to do mental calculations on differences between slices.

Patch-set: 2
CC: Gerrit User 5310 <5310@62eb7196-b449-3ce5-99f1-c037f21e1705>