Update patch set 4
Patch Set 4:

I see.  Thanks for the clarification.  But I'm still not comfortable with this.  As you know the linker is very complicated, because it is full of different conditions for when to take certain actions.  This adds another such conditional.  It's pretty minor, but if the only purpose is to speed up the linker, then I would prefer to keep the slightly simpler code, rather than make people try to understand what is happening.  And if the change has some purpose other than speeding up the linker, I don't understand what it is.

Patch-set: 4