Update patch set 4
Patch Set 4:

> > I see.  Thanks for the clarification.  But I'm still not comfortable with this.  As you know the linker is very complicated, because it is full of different conditions for when to take certain actions.  This adds another such conditional.  It's pretty minor, but if the only purpose is to speed up the linker, then I would prefer to keep the slightly simpler code, rather than make people try to understand what is happening.  And if the change has some purpose other than speeding up the linker, I don't understand what it is.
> 
> Actually, my purpose is just simplifying/clarifying code. If you think this CL make the situation more complicated, I failed.
> 
> I'd like to remove some kludges by introducing the precondition -- dynreloc is not intended to work in external linking mode.

From my perspective, every time we write `if linkmode == LinkExternal` we are making the linker more complex, because now we have to understand why it works differently.  It's better to have one code path than two.

It would be simpler to write
    if linkmode == LinkExternal {
        log.Fatal("should not have gotten here")
    }
That is nice and clear.  But it is not simpler to write
    if linkmode == LinkExternal {
        return
    }
because now whenever we change anything that might change what `dynreloc` does, we have to think about whether it really does change it, and what the link mode is for that change.

Patch-set: 4