Update patch set 1
Patch Set 1:

> Patch Set 1:
> Daniel, is this something that unused could/would check for?

Assuming you mean unparam - at the moment not, by "parameters" it understands received parameters, not returning parameters.

Potentially it could, but not sure if it should. It could make for quite a bit of added complexity, but it's hard to tell without giving it a go. Care to open an issue about it?

Patch-set: 1