Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why split declaration and assignment?
> No reuse, I would like to understand why it is better to make declaration and assignment
> 
> About implementing this correct

As I said in the commit message, in this case, it improves the generated code. The reason is tied up with the timing of a compiler pass; see the issue I linked to in the commit message.

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>