Update patch set 1
Patch Set 1:

I'm not sure the CL description is precisely accurate.  It seems to me that we were already protected against (newcap * et.size) overflow.  If this adds any protection, it is against roundupsize overflow.  For that case I think it changes the failure mode from a throw to a panic, which I guess is a good thing.  But are there really any sizes where newcap * et.size does not overflow but roundupsize does overflow?

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>