Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > I see where you've added utimensat(), but I don't see where you're actually calling it, I would have expcted UtimesNano and UtimesNanoAt to be updated to use it, same as I did here:
> > 
> > https://github.com/golang/sys/blob/master/unix/syscall_solaris.go#L304
> 
> There is a patch to syscall_bsd.go to call utimensat from UtimesNano.  Note that this is to the syscall package, not the golang.org/x/sys/unix package.

I'm aware that this is a patch to the syscall package, but this changeset claims to fix both *BSD and Solaris.

It only fixes things for BSD, I do not see any change to go/syscall_solaris.go to change UtimesNano to call utimensat().  Could you please point me to the change I somehow missed?

Patch-set: 2