Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Note that this CL is a followup to golang.org/cl/53471 which changes an existing check to capmem > _MaxMem and is included here.
> > 
> > I do not see where without this CL we check that the multiplication does not overflow. We do not check newcap against something with _MaxMem or maxUint and et.Size before we use it in roundupsize.
> > 
> > MaxSliceCap implicitly provides a maximum cap that would not overflow when multiplied by et.size since AFAIK _MaxMem <= maxUint.
> > 
> > Since the new overflow var is not directly dependent on capmem it can not be checking the roundupsize overflow.
> > 
> > I have to check again if roundsize can overflow but it think roundupsize either rounds to sizeclasses which are < _MaxMem which is safe. Or to largeAlloc and in that case the next power of 2 which depending on _MaxMem size could pose a problem. Will check that.
> 
> I'm sure you're right, but it would be helpful if you could give an example of a case that currently overflows.

Fair question. I will try to force a failure. Might actually be impossible to allocate enough memory in makeslice for a trigger on 386 to begin with. I could resort to using unsafe. While its then is practically not possible without unsafe, it might be good to document we rely on some constraints outside of growslice, so that we do not accidentally introduce the possibility of overflow later e.g. if we ever support append with multiple arguments that are appended: "append(x,a...,b...)".

I will work on a fix for the newcap infinite loop first before getting back to this CL.

Patch-set: 1
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>