Update patch set 1
Patch Set 1:

I had noticed the missing 'int flags' parameter in the original change request, but had (foolishly) assumed that was intentional (and ok) because of the calling mechanism and because syscall_linux.go was already that way.

This change matches what was done in sys/unix.

With that in mind, should syscall_linux.go be updated to match sys/unix too and add the 'flags' parameter?

Patch-set: 1
CC: Gerrit User 7291 <7291@62eb7196-b449-3ce5-99f1-c037f21e1705>