Update patch set 3
Patch Set 3:

@Dave

i had a look and these build breakages due to submit order i think are not frequent enough to warrant needing a rebase if the CL is not at current master head. The breakage might not be related to the current package but might also happen due to a change in another package: golang.org/cl/38776

So the only way to make sure it does not break the build is to test it like the builder does or at least only test all dependencies as the builder does (which would be everything anyway in case of runtime).

If you think this warrants further discussion to find some solution i am happy to open a github issue for it.

Patch-set: 3