cmd/asm: uncomment tests for already implemented amd64 SSE4 instructions
Uncomments PCMPESTRI and PHMINPOSUW test cases.

It is unknown when x86test will be updated/re-run, but tests are useful
to check which x86 instructions are not yet supported.
As an example of tool that uses this information, there is Damien
Lespiau x86db.

Part of the mission to add missing amd64 SSE4 instructions to Go asm.

Change-Id: I512ff26040f47a0976b3e37000fb1f37eac5b762