runtime,testing: remove some unused result params
In assertE2I2, we set up an iface variable but it was never used. Avoid
that work. While at it, simplify the return logic since it's just a
bool.

And in B.run, the result was never used. Nor is it necessary; it's a
field and can still be accessed easily.

Found with github.com/mvdan/unparam, after Kevin Burke's suggestion that
the tool should also warn about unused result parameters.

Change-Id: Id8b5ed89912a99db22027703a88bd94d0b292b8b