Update patch set 3
Patch Set 3:

I don't understand why this change is so large. Why wasn't this simply adding a few entries to the existing generator.

Also, the generator code now seems awfully complicated and long. In fact the generator code size is almost the same size as the generated code and it's by far more complex.

I think this defeats the purpose. Surely there's a simpler way to approach this. If not, we might just as well throw away the generator and just keep and extend the latest generated code.

Patch-set: 3
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>