cmd/compile: loop preemption with "fault branch" on amd64
Loop preemption turned on by default, not by GOEXPERIMENT
Intended for handoff to Austin for runtime fiddling.

On amd64, "preempt test" is a load from runtime.reschedulePage
On non-amd64, the prologue test is repeated because the
probe is not yet implemented on other platforms and because
page size issues make the current probe declaration not
portable.

To force probe-test:
-gcflags=-d=ssa/preeempt_loops/test=2

To force stackbound-test:
-gcflags=-d=ssa/preeempt_loops/test=1

To disable loop preemption for a build:
-gcflags=-d=ssa/preeempt_loops/off

TODO: this causes a regression in loopbce.go (test stubbed out)
TODO: runtime half of probe-based test is missing.
TODO: how is "target" of BlockIfFault communicated to runtime?
TODO: register allocation is slightly degraded

Updates: #10958

Change-Id: I9265a20b6a95b441a4794f7cedbb25e30dd88b2a