Update patch set 1
Patch Set 1:

Personally I don't find these entirely convincing.  I think it's normal to write

if x == 1 {
} else if x == 2 {
}

but I think it's odd to write

if x == 1 {
} else if y == 2 {
}

When I see an if/else if/ I normally expect the condition in the second if to have some relationship to the condition in the first if.  In these examples that isn't there.  So it doesn't seem so bad to write

if x == 1 {
} else {
    if y == 2 {
    }
}

It could be written in fewer lines, yes, but pushing into a new block calls out that the second condition is a different kind of test.

I'm open to counter-arguments.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>