Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> CL/56771 adds documentation for SparseHoles. Was that not sufficient for you?

That's just a reference; it's moderately good for reading sparse files (though you also need to find the mention in Read to fully understand what is going on), but not for writing.  It doesn't tell me how to create a sparse file. Should I expand it (writing NUL) while writing? What happens if I set SparseHole but not typeflag? Is there any way to copy a sparse file from disk into a tar without expanding it?

I think there are many questions to be answered, and the chosen API is not very clear. It should at least be clearly documented. I see that archive/tar has basically none package-level documentation, so I don't have good suggestions on where to write a couple of paragraph on how to handle sparse files.

Patch-set: 1