Update patch set 2
Patch Set 2: Code-Review-1

The updates to the documentation looks good. The package still fails to give a clear answer the following questions:

 * If I have a tar file that contains a sparse file, how can I extract it to disk (keeping it sparse)?
 * If I have a sparse file on disk, how can I write it into the tar file (as sparse file)?

And this is why I call the API "not clear": it only provides some low-level bricks (SparseHole); so it certainly make sparse file support "technically possible" but in a complex way that requires writing lots of careful code on top of it, and it's not very clear how to do it. If the API was good, it would probably require almost no documentation to do the correct thing.

So I don't think this bug "fixes #13548", at most it "updates #13548" by adding some low-level primitive. I don't think this should go in until the final API (which allows people to handle sparse file in a simple way) is designed.

Patch-set: 2
Reviewer: Gerrit User 5340 <5340@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1