Update patch set 2
Patch Set 2:

(2 comments)

Making the test high level makes sense. However, I wonder if there's a decent way to move it to flag package and keep the test behavior the same. That'd make it so that regressions to flag.Fatalf behavior are caught by the person running `go test flag` more quickly.

Patch-set: 2