Update patch set 2
Patch Set 2:

>  * If I have a tar file that contains a sparse file, how can I extract it to disk (keeping it sparse)?
>  * If I have a sparse file on disk, how can I write it into the tar file (as sparse file)?

There's several distinct issues at hand:
1. Support for encoding sparse files in TAR format (which this CL addresses)
2. Support for easily writing/reading sparse files to/from disk (this involves either adding Reader.Discard/Writer.FillZeros or Reader.WriteTo/Writer.ReadFrom or both sets of them).
3. Support for detecting sparse holes in files on disk (this may involve adding a FileHeader that populates a Header from an os.File).

Your concerns are related to #2 and #3. They are good concerns, but they are beyond the scope of what this CL is trying to accomplish. Unless you have a concern about the actual implementation of sparse file in the TAR format, I think it would be best to address those two concerns separately.

> So I don't think this bug "fixes #13548", at most it "updates #13548" by adding some low-level primitive. I don't think this should go in until the final API (which allows people to handle sparse file in a simple way) is designed.

That's fine. I changed it to Updates #13548. However, I'm not in favor of adding sparse-file support to Writer without exposing that information to the user. In which case, this CL is still an intermediate step that we need to go through to get to where we want to be.

Patch-set: 2