Update patch set 2
Patch Set 2:

> I'm open to suggestions. I of course thought about putting it in the flag package first, but I have no idea how to unit test a function that calls os.Exit unconditionally.

I'll try to take a look at it.

Basically, I'm imagining something that calls https://godoc.org/internal/testenv#HasGoBuild first, then writes a .go file to a temp dir, go builds it, then executes it with os/exec and verifies that the exit code of the process is 2. The main question is looking if there's some existing utilities available to write less new code to make that happen.

Patch-set: 2