spec: explicitly define notion of "representability" (clarification)
Throughout the spec we use the notion of a constant x being
representable as a value of type T. While intuitively clear,
at least for floating-point and complex constants types, the
concept was not well-defined. In the section on Conversions
there was an extra rule for floating-point types only.

Since the concept is important to Go, and a compiler most
certainly will have a function to test "representability",
it seems warranted to define the term explicitly in the spec.

This change introduces a new entry "Representability" under
the section on "Properties of types and values", and defines
the term explicitly. It also introduces links throughout the
spec to this section when the term is used. Finally, it removes
the special case for floating-point constant conversions. As
as result, the section on constant conversions can be simplified.

Fixes #15389.

Change-Id: I8be0e071552df0f18998ef4c5ef521f64ffe8c44