Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> >  * If I have a tar file that contains a sparse file, how can I extract it to disk (keeping it sparse)?
> >  * If I have a sparse file on disk, how can I write it into the tar file (as sparse file)?
> 
> There's several distinct issues at hand:
> 1. Support for encoding sparse files in TAR format (which this CL addresses)
> 2. Support for easily writing/reading sparse files to/from disk (this involves either adding Reader.Discard/Writer.FillZeros or Reader.WriteTo/Writer.ReadFrom or both sets of them).
> 3. Support for detecting sparse holes in files on disk (this may involve adding a FileHeader that populates a Header from an os.File).
> 
> Your concerns are related to #2 and #3. They are good concerns, but they are beyond the scope of what this CL is trying to accomplish. Unless you have a concern about the actual implementation of sparse file in the TAR format, I think it would be best to address those two concerns separately.

What this CL does is to add some low-level support for encoding sparse files, *and* add a "low-level API" (if you agree with this naming) which I consider insufficient by itself; on top of this, it fails to provide guidance (even if only at the documentation level) on how to use the low-level API to achieve full sparse file support.

The disagreement between us is whether we should go forward with this low-level API before having figured out how the final "high-level API" would look like. For instance, in the issue I made a high-level API proposal which conflicts with your low-level API (though you seem to be against it). I feel that, until a proposal for the high-level API is finalized and agreed upon, this new low-level API should not be released, because we could then have to maintain it forever even if it turns out to be the wrong thing to do. This is why I put my -1 here.

Patch-set: 3
Reviewer: Gerrit User 5340 <5340@62eb7196-b449-3ce5-99f1-c037f21e1705>