runtime: only clear pointer-containing memory during map delete
When deleting entries from a map, only clear the key and value
if they contain pointers. This will be a slight regression for
keys and values with pointers and a significant speed-up for
keys and values without pointers.

While we're here, specialize key clearing in mapdelete_faststr.

Benchmarking impeded by #21546.

Change-Id: I3f6f924f738d6b899b722d6438e9e63f52359b84